+++ src/components/mypage/MyPosts/index.tsx
@@ -11,8 +11,9 @@ import {
   formatCategoryToGameNameKR,
   formatGameToLink,
   splitTitleByDelimiter,
+  getPostPageSize,
+  getProcessedDataList,
 } from '@/utils';
-import { getPostPageSize } from '@/utils/getPageSize';
 
 import { RegisteredCard } from '@/components/commons/cards';
 import { CardSkeleton } from '@/components/commons/cards/CardSkeleton';
@@ -21,7 +22,6 @@ import Filter from '@/components/commons/Filter';
 import Pagination from '@/components/commons/Pagination';
 import EmptyCard from '@/components/layout/empty/EmptyCard';
 import { useDeviceType } from '@/hooks/useDeviceType';
-import useProcessedDataList from '@/hooks/useProcessedDataList';
 
 import { MyActivitiesResponse, Order, SortOption } from '@/types';
 
@@ -52,12 +52,11 @@ const MyPosts = () => {
   const currentDeviceType = useDeviceType();
   const pageSize = getPostPageSize(currentDeviceType);
 
-  const { pagedDataList, totalCount } = useProcessedDataList({
+  const { pagedDataList, totalCount } = getProcessedDataList({
     initialDataList,
     selectFilter,
     sortOption,
     page,
-    setPage,
     postsPerPage: pageSize,
   });
