+++ src/components/mypage/MyReservations/ReservedTabContent/index.tsx
@@ -6,8 +6,7 @@ import classNames from 'classnames/bind';
 import { getMyReservations } from '@/apis/queryFunctions';
 import { QUERY_KEYS } from '@/apis/queryKeys';
 import { GAME_NAME_KR_TO_PATH_NAME, MY_RESERVATIONS_STATUS_FILTERS, SORT_OPTIONS } from '@/constants';
-import { formatStatusToKR, splitTitleByDelimiter } from '@/utils';
-import { getPostPageSize } from '@/utils/getPageSize';
+import { formatStatusToKR, splitTitleByDelimiter, getPostPageSize, getProcessedDataList } from '@/utils';
 
 import { ReservedCard } from '@/components/commons/cards';
 import { CardSkeleton } from '@/components/commons/cards/CardSkeleton';
@@ -16,7 +15,6 @@ import Filter from '@/components/commons/Filter';
 import Pagination from '@/components/commons/Pagination';
 import EmptyCard from '@/components/layout/empty/EmptyCard';
 import { useDeviceType } from '@/hooks/useDeviceType';
-import useProcessedDataList from '@/hooks/useProcessedDataList';
 
 import { ReservationResponse, SortOption, Order } from '@/types';
 
@@ -47,12 +45,11 @@ const ReservedTabContent = () => {
   const currentDeviceType = useDeviceType();
   const pageSize = getPostPageSize(currentDeviceType);
 
-  const { pagedDataList: reservationData, totalCount } = useProcessedDataList({
+  const { pagedDataList: reservationData, totalCount } = getProcessedDataList({
     initialDataList,
     selectFilter,
     sortOption,
     page,
-    setPage,
     postsPerPage: pageSize,
   });
